Ports from SDL2 to SDL3
Sets the default C version to C99

Re-worked controller handling
    The controller handling has been improved to allow for a more
    seamless connect/disconnect experience. Controllers are now loaded
    (if rezognized) when they are connected rather then at game boot.

    The tooltip system has been re-written to allow for loading 3 types
    of tooltips in one sitting. One for each controller type. The cost
    is some extra sprite memory. But in all fairness this is not a
    massiv impact on the game.

    This change introduced the `typedef struct tooltip Tooltip` as well
    as `enum GamepadType`.

+++ src/timer.h
@@ -20,16 +20,17 @@
 #define	TIMER_H_
 
 #include <stdbool.h>
+#include <SDL3/SDL.h>
 
 typedef struct Timer {
-	unsigned int startTime;
+	Uint64 startTime;
 } Timer;
 
 Timer* _timer_create(void);
 void timer_start(Timer*);
 void timer_stop(Timer*);
 bool timer_started(Timer*);
-unsigned int timer_get_ticks(Timer*);
+Uint64 timer_get_ticks(Timer*);
 void timer_destroy(Timer*);
 
 #endif // TIMER_H_
